Devnet updates: Genesis with CW XCVM contracts for Substrate devnet and Osmosisd app (#3737)
- CW in CI green and builds all the time
- Osmosisd available
- CW in genesis of Picasso (some issue with interpreter contract in
genesis)
- Centauri chain update to latest version and process-compose update
- relayer of Centauri <-> Osmosis

next:
- centauiry hack for pica denom
- CW genesis in Osmosis/Centauri, init, start
- Memo limit for Osmosis 2x increase hack
- composable node with isolation for p2p discovery and 3rd relayer
- macos darwin support (seem couple of lines + build in CI on main)
- starting from prebaked genesis




- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [ ] I have linked Zenhub/Github or any other reference item if one
exists
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [x] I waited and did best effort for `pr-workflow-check /
draft-release-check` to finish with success(green check mark) with my
changes
- [x] I have added at least one reviewer in reviewers list
- [x] I tagged(@) or used other form of notification of one person who I
think can handle best review of this PR
- [x] I have proved that PR has no general regressions of relevant
features and processes required to release into production

---------

+++ inputs/notional-labs/composable-centauri/flake-module.nix
@@ -2,6 +2,7 @@
   perSystem = { config, self', inputs', pkgs, lib, system, crane
     , systemCommonRust, subnix, ... }:
     let
+      devnet-root-directory = "/tmp/composable-devnet";
       validator = "centauri12smx2wdlyttvyzvzg54y2vnqwq2qjatescq89n";
       validator-mnemonic =
         "bottom loan skill merry east cradle onion journey palm apology verb edit desert impose absurd oil bubble sweet glove shallow size build burst effort";
@@ -18,13 +19,13 @@
         excludedPackages = [ "interchaintest" "simd" ];
         ldflags = [ "-v -extldflags '-L${self'.packages.libwasmvm}/lib'" ];
         src = pkgs.fetchFromGitHub {
-          owner = "notional-labs";
+          owner = "dzmitry-lahoda-forks";
           repo = "composable-centauri";
-          rev = "762ad8efd57695ff4ee469aa6923fe7e81578c61";
-          sha256 = "sha256-/wN246AZ4eO9Z2+q7iLLw33G+7cbUCCtetA2BFDRlj8=";
+          rev = "c677c9f121b2637f0e9cca370a8fa0f72a2298c9";
+          sha256 = "sha256-clAuk5UrSeRooLiqFPcSk6r6TkBe3tUIn4bXrS24IPk=";
         };
         dontFixup = true;
-        vendorSha256 = "sha256-xbHVzucr/5B9ePP3Dnhag9KaFKvS238sTgkoxsD3LJ8=";
+        vendorSha256 = "sha256-2isPCn91vSeNmMw58mu0yP7UGDXHNJW+6K2aNh2dKJc=";
       };
 
       ibc-lightclients-wasm-v1-msg-push-new-wasm-code = code: {
@@ -76,30 +77,32 @@
         runtimeInputs = [ centaurid pkgs.jq pkgs.yq ];
 
         text = ''
-          CENTAURI_DATA="/tmp/centauri-dev"
+          CENTAURI_DATA="${devnet-root-directory}/.centaurid"
           CHAIN_ID="centauri-dev"
           KEYRING_TEST="$CENTAURI_DATA/keyring-test"
-          # centaurid query bank balances ${validator} --chain-id "$CHAIN_ID" --node tcp://localhost:26657 --home "$CENTAURI_DATA"
-          # centaurid query bank balances centauri1cyyzpxplxdzkeea7kwsydadg87357qnamvg3y3 --chain-id "$CHAIN_ID" --node tcp://localhost:26657 --home "$CENTAURI_DATA"
-          # centaurid tx bank send validator centauri1cyyzpxplxdzkeea7kwsydadg87357qnamvg3y3 1000000000ppica --from validator --keyring-backend test --gas 902152622 --fees 920166ppica --keyring-dir "$KEYRING_TEST" --chain-id "$CHAIN_ID" --yes --output json
-          # sleep 5
-          # centaurid query bank balances ${validator} --chain-id "$CHAIN_ID" --node tcp://localhost:26657 --home "$CENTAURI_DATA"
-          # centaurid query bank balances centauri1cyyzpxplxdzkeea7kwsydadg87357qnamvg3y3 --chain-id "$CHAIN_ID" --node tcp://localhost:26657 --home "$CENTAURI_DATA"            
-
-          echo "=============== SUBMIT PROPOSAL ========"
           centaurid tx gov submit-proposal ${ics10-grandpa-cw-proposal}/ics10_grandpa_cw.wasm.json --from "${validator}"  --keyring-backend test --gas 9021526220000 --fees 92000000166ppica --keyring-dir "$KEYRING_TEST" --chain-id "$CHAIN_ID" --yes --home "$CENTAURI_DATA" --output json
-          sleep 5          
-
-          echo "=============== VOTE PROPOSAL ========="
-          PROPOSAL_ID=1
-          centaurid tx gov vote $PROPOSAL_ID yes --from "${validator}"  --keyring-backend test --gas 9021526220000 --fees 92000000166ppica --keyring-dir "$KEYRING_TEST" --chain-id "$CHAIN_ID" --yes --home "$CENTAURI_DATA" --output json
-          sleep 20
-          echo "=============== GET PROPOSAL ========="
+          sleep 5
+          centaurid query auth module-account gov --chain-id "$CHAIN_ID" --node tcp://localhost:26657 --home "$CENTAURI_DATA" | jq '.account.base_account.address' --raw-output
+          PROPOSAL_ID=1          
+          centaurid tx gov vote $PROPOSAL_ID yes --from "${validator}"  --keyring-backend test --gas 9021526220000 --fees 92000000166ppica --keyring-dir "$KEYRING_TEST" --chain-id "$CHAIN_ID" --yes --home "$CENTAURI_DATA" --output json          
+          sleep 20          
           centaurid query gov proposal $PROPOSAL_ID --chain-id "$CHAIN_ID" --node tcp://localhost:26657 --home "$CENTAURI_DATA" |
           jq '.status'
-          sleep 5
-          echo "=============== GET WASM ========="
+          sleep 5          
           centaurid query 08-wasm all-wasm-code --chain-id "$CHAIN_ID" --home "$CENTAURI_DATA" --output json --node tcp://localhost:26657 | jq '.code_ids[0]' --raw-output | tee "$CENTAURI_DATA/code_id"
+        '';
+      };
+
+      centaurid-dev = pkgs.writeShellApplication {
+        name = "centaurid-dev";
+        runtimeInputs = [ centaurid pkgs.jq pkgs.yq ];
+
+        text = ''
+          #CENTAURI_DATA="${devnet-root-directory}/.centaurid"
+          CHAIN_ID="centauri-dev"
+          #KEYRING_TEST="$CENTAURI_DATA/keyring-test"
+          centaurid query ibc connection connections --chain-id "$CHAIN_ID"
+          #--from "${validator}"  --keyring-backend test --gas 9021526220000 --fees 92000000166ppica --keyring-dir "$KEYRING_TEST"  --yes --home "$CENTAURI_DATA" --output json
         '';
       };
 
@@ -107,10 +110,14 @@
         name = "centaurid-gen";
         runtimeInputs = [ centaurid pkgs.jq pkgs.yq ];
         text = ''
-          CENTAURI_DATA="/tmp/centauri-dev"
+          CENTAURI_DATA="${devnet-root-directory}/.centaurid"
           CHAIN_ID="centauri-dev"
           KEYRING_TEST="$CENTAURI_DATA/keyring-test"
-          rm --force --recursive "$CENTAURI_DATA"
+          REUSE=true
+          export REUSE
+          if [[ $REUSE == false ]]; then
+            rm --force --recursive "$CENTAURI_DATA" 
+          fi
 
           mkdir --parents "$CENTAURI_DATA"
           mkdir --parents "$CENTAURI_DATA/config/gentx"
@@ -145,9 +152,12 @@
           echo "notice oak worry limit wrap speak medal online prefer cluster roof addict wrist behave treat actual wasp year salad speed social layer crew genius" | centaurid keys add test1 --recover --keyring-backend test --keyring-dir "$KEYRING_TEST" || true
           echo "quality vacuum heart guard buzz spike sight swarm shove special gym robust assume sudden deposit grid alcohol choice devote leader tilt noodle tide penalty" | centaurid keys add test2 --recover --keyring-backend test --keyring-dir "$KEYRING_TEST" || true
           echo "symbol force gallery make bulk round subway violin worry mixture penalty kingdom boring survey tool fringe patrol sausage hard admit remember broken alien absorb" | centaurid keys add test3 --recover --keyring-backend test --keyring-dir "$KEYRING_TEST" || true
+          echo "black frequent sponsor nice claim rally hunt suit parent size stumble expire forest avocado mistake agree trend witness lounge shiver image smoke stool chicken" | centaurid keys add relayer --recover --keyring-backend test --keyring-dir "$KEYRING_TEST" || true
           function add-genesis-account () {
             centaurid --keyring-backend test add-genesis-account "$1" "1000000000000000000000ppica" --keyring-backend test --home "$CENTAURI_DATA"          
           }
+
+          add-genesis-account centauri1qvdeu4x34rapp3wc8fym5g4wu343mswxxgc6wf
           add-genesis-account centauri1zr4ng42laatyh9zx238n20r74spcrlct6jsqaw
           add-genesis-account centauri1makf5hslxqxzl29uyeyyddf89ff7edxyr7ewm5
           add-genesis-account ${validator}
@@ -156,12 +166,12 @@
           add-genesis-account centauri1qwexv7c6sm95lwhzn9027vyu2ccneaqapystyu
           centaurid --keyring-backend test --keyring-dir "$KEYRING_TEST" --home "$CENTAURI_DATA" gentx validator "250000000000000ppica" --chain-id="$CHAIN_ID" --amount="250000000000000ppica"
           centaurid collect-gentxs --home "$CENTAURI_DATA"  --gentx-dir "$CENTAURI_DATA/config/gentx"
-          centaurid start --rpc.unsafe --rpc.laddr tcp://0.0.0.0:26657 --pruning=nothing  --minimum-gas-prices=0ppica --log_level debug --home "$CENTAURI_DATA" --db_dir "$CENTAURI_DATA/data" --log_format json --trace --with-tendermint true --transport socket --trace-store $CENTAURI_DATA/kvstore.log --grpc.address localhost:9090 --grpc.enable true --grpc-web.enable false --api.enable true --cpu-profile $CENTAURI_DATA/cpu-profile.log
+          centaurid start --rpc.unsafe --rpc.laddr tcp://0.0.0.0:26657 --pruning=nothing  --minimum-gas-prices=0ppica --log_level debug --home "$CENTAURI_DATA" --db_dir "$CENTAURI_DATA/data" --trace --with-tendermint true --transport socket --trace-store $CENTAURI_DATA/kvstore.log --grpc.address localhost:9090 --grpc.enable true --grpc-web.enable false --api.enable true --cpu-profile $CENTAURI_DATA/cpu-profile.log --p2p.pex false --p2p.upnp  false
         '';
       };
     in {
       packages = rec {
-        inherit centaurid centaurid-gen centaurid-init
+        inherit centaurid centaurid-gen centaurid-init centaurid-dev
           ics10-grandpa-cw-proposal;
       };
     };
